OPENNLP-1663 Add test for FileToByteArraySampleStream - adds FileToByteArraySampleStreamTest - moves FileToStringSampleStreamTest to the correct package
