+++ opennlp-tools/src/test/java/opennlp/tools/formats/convert/AbstractConvertTest.java
@@ -15,10 +15,9 @@
  * limitations under the License.
  */
 
-package opennlp.tools.convert;
+package opennlp.tools.formats.convert;
 
 import java.io.IOException;
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -26,41 +25,28 @@ import java.nio.file.StandardOpenOption;
 import java.util.Arrays;
 import java.util.List;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeEach;
 
 import opennlp.tools.AbstractTempDirTest;
 import opennlp.tools.formats.DirectorySampleStream;
-import opennlp.tools.formats.convert.FileToStringSampleStream;
 
-public class FileToStringSampleStreamTest extends AbstractTempDirTest {
+public abstract class AbstractConvertTest extends AbstractTempDirTest {
 
-  @Test
-  public void readFileTest() throws IOException {
+  protected DirectorySampleStream sampleStream;
+  protected List<String> sentences;
 
+  @BeforeEach
+  public void setUp() throws IOException {
     final String sentence1 = "This is a sentence.";
     final String sentence2 = "This is another sentence.";
 
-    List<String> sentences = Arrays.asList(sentence1, sentence2);
-
-    DirectorySampleStream directorySampleStream =
-        new DirectorySampleStream(tempDir.toFile(), null, false);
+    sentences = Arrays.asList(sentence1, sentence2);
+    sampleStream = new DirectorySampleStream(tempDir.toFile(), null, false);
 
     Path tempFile1 = tempDir.resolve("tempFile1");
     Files.writeString(tempFile1, sentence1, StandardCharsets.UTF_8, StandardOpenOption.CREATE);
 
     Path tempFile2 = tempDir.resolve("tempFile2");
     Files.writeString(tempFile2, sentence2, StandardCharsets.UTF_8, StandardOpenOption.CREATE);
-
-    try (FileToStringSampleStream stream =
-             new FileToStringSampleStream(directorySampleStream, Charset.defaultCharset())) {
-
-      String read = stream.read();
-      Assertions.assertTrue(sentences.contains(read));
-
-      read = stream.read();
-      Assertions.assertTrue(sentences.contains(read));
-    }
   }
-
 }
